RecipeFormViewModel unit tests

+++ app/src/main/java/com/aamo/cookbook/features/recipe/form/use_cases/DeleteRecipe.kt
@@ -2,13 +2,14 @@ package com.aamo.cookbook.features.recipe.form.use_cases
 
 import com.aamo.cookbook.database.entities.Recipe
 
-// TODO: unit test
 suspend fun deleteRecipe(
   recipe: Recipe,
   deleteThumbnail: suspend (String) -> Unit,
   deleteRecipe: suspend (Recipe) -> Boolean,
 ): Boolean {
-  return deleteRecipe(recipe).also {
+  if (deleteRecipe(recipe)) {
     if (recipe.thumbnailUri.isNotEmpty()) deleteThumbnail(recipe.thumbnailUri)
+    return true
   }
+  return false
 }
